iip_config/requirements.txt
django-markdown-deux==1.0.5
Django==1.4.10
envoy==0.0.2
importlib==1.0.3
lxml==3.3.0
mysolr==0.8
requests==2.2.1
rq-dashboard==0.3.3
rq==0.3.13
shellvars-py==0.1.2
solrpy==0.9.6

django-crispy-forms==1.6.1
django-pagedown==0.1.1
